requirements.txt
icecream==2.1.3
openai==1.35.3
pdf2image==1.17.0
Pillow==10.3.0
python-dotenv==1.0.1
